Tentative finish to file parser script
Ideally this is what we want the file parser to look like, with each line
type handled correctly by the action in parseme. This has not yet been
tested (I'm leaving for my cousin's now) but leave me a comment in the pull
request on the last two commits I've made?